Add files via upload
This is a comprehensive analysis of Glitz-Auto car sales for the year 2022-2023 showcasing the trends and pattern of the business strategy to provide insights on top performing car dealer, the regions in which the highest sales were realized from, including best performing car companies. Down below are my observations and recommendations:

OBSERVATIONS:

From the sales trend report, i observed that in year 2023, $10,749,111,884 is the total sales accumulated which surpass year 2022's total sales of $9,112,955,972. So after summing the values, the company got a whopping sum of $19,862,067,856 as total revenue. the top region is Austin, $3,402,482,981 sales was realized from that region, making Austin the region with the highest sales. 

RECOMMENDATIONS:

Though other regions did well, but the difference between the revenue gotten from other regions when compared to the revenue realized from the top performing region is not much. having said that, a new and better market strategy should be employed and more competent sales person and workers should be employed as well to enhance the operation.

OBSERVATION:

From the data, the top performing companies are; Ford, Chevrolet, Volkswagen and Dodge and the leading company is Chevrolet with $1,453,860,658 total sales recorded. In Janesville region, chevrolet cars were sold more there and also Star Enterprises Inc (car dealer) recorded the highest sales.

RECOMMENDATIONS: 

Progressive shippers association which is the top car dealer recorded a low sale of Chevrolet cars which is the most sold car. This could be as a result of inadequate supply of Chevrolet cars or poor advertisements of the car. I would recommend that better adverts should be done and also there should be high supply of the product to match the demands for it. 

OBSERVATIONS:

The top performing region is Austin. $3,402,482,981 sales was realized from that region, making Austin the region with the highest sales. In this region, the top dealer is Race Car Help, $870,409,898 was accumulated in that region.

RECOMMENDATIONS:

It appears that in Austin people buy race cars more and less of used cars, so more racing cars should be made available in that region with motro sport, LS400 or pathfinder. Supply of used cars to this region should be redirected to the regions where they are in high demand such as, Scottdale.